Modify YoutubeComponent structure for toggle btn
